uploader fine tunning
cancel and delete button together
fix retry
implementing more intelligent retry algorithm
title on the bottom, progress on the top

+++ app/commands/destroy_upload.rb
@@ -8,6 +8,7 @@ class DestroyUpload
 
   def call
     upload = Upload.find_by_uuid(upload_attributes[:uuid])
+    return publish :not_found unless upload
 
     if upload.destroy
       cloud_storage_object.new(upload.object_key).delete
